backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40",
    "beir>=2.0",
    "fastapi>=0.128.0",
    "numpy>=2.0",
    "openai>=1.50",
    "orjson>=3.11.5",
    "pydantic-settings>=2.12.0",
    "pymupdf==1.26.6",
    "pymupdf4llm>=0.0.17",
    "pymupdf_layout==1.26.6",
    "python-multipart>=0.0.21",
    "rapidfuzz>=3.14.3",
    "rich>=14.2.0",
    "sentence-transformers>=3.0",
    "tavily-python>=0.5",
    "typer>=0.21.1",
    "uvicorn[standard]>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/metis"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
metis = "metis.adapters.cli:main"
metis-web = "metis.adapters.web:main"
